# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "SpotSync Parking Management API",
        "title": "SpotSync API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "SpotSync Team",
            "email": "support@spotsync.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/zones": {
            "get": {
                "description": "Retrieve all parking zones",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parking Zones"
                ],
                "summary": "Get All Parking Zones",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Success"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new parking zone (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parking Zones"
                ],
                "summary": "Create Parking Zone",
                "parameters": [
                    {
                        "description": "Parking Zone",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sportsync-api_internal_domain_admin_dto.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    }
                }
            }
        },
        "/zones/{id}": {
            "get": {
                "description": "Retrieve a parking zone by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parking Zones"
                ],
                "summary": "Get Parking Zone",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parking Zone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing parking zone (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parking Zones"
                ],
                "summary": "Update Parking Zone",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parking Zone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Parking Zone",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sportsync-api_internal_domain_admin_dto.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a parking zone (Admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Parking Zones"
                ],
                "summary": "Delete Parking Zone",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Parking Zone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Success"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpResponse.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "httpResponse.Error": {
            "type": "object",
            "properties": {
                "errors": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "httpResponse.Success": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "sportsync-api_internal_domain_admin_dto.CreateRequest": {
            "type": "object",
            "required": [
                "name",
                "price_per_hour",
                "total_capacity",
                "type"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "price_per_hour": {
                    "type": "number"
                },
                "total_capacity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "general",
                        "ev_charging",
                        "covered"
                    ]
                }
            }
        },
        "sportsync-api_internal_domain_admin_dto.UpdateRequest": {
            "type": "object",
            "required": [
                "name",
                "price_per_hour",
                "total_capacity",
                "type"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "price_per_hour": {
                    "type": "number"
                },
                "total_capacity": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "general",
                        "ev_charging",
                        "covered"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}